+++ src/consts.js
@@ -1,4 +1,3 @@
-// NOTE: This is not nice, it waits for implementing default timeout into puppeteer.
 exports.DEFAULT_TIMEOUT = 60 * 1000; // 60 sec
 
 exports.LISTING_PAGINATION_KEY = 'lisState';